+++ NEWS.md
@@ -1,3 +1,5 @@
+* Fixing bug in *quick_map* occurring when *plot_col* is identical to one of the columns in the table *countries::world*
+
 # countries 1.1.1
 * Added function *list_countries* to obtain a list of countries
 * Added function *random_countries* to obtain random country names